src/requirements.txt
beautifulsoup4==4.14.3
requests==2.32.5
Pillow==12.1.0
boto3==1.42.44
ruff==0.15.0
pyrefly==0.51.2
